+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog for package layered_hardware_gazebo
 
 0.0.5 (next version)
 --------------------
+* Add mimic/mimic-PID operation modes
 * Supress a warning on unloading layer plugins
 
 0.0.4 (2020-05-05)
@@ -19,7 +20,7 @@ Changelog for package layered_hardware_gazebo
 
 0.0.2 (2020-04-30)
 ------------------
-* Add pisition/velocity/posvel PID operation mode (effort-based position/velocity/posvel mode)
+* Add pisition/velocity/posvel -PID operation modes (effort-based position/velocity/posvel mode)
 * Add posvel operation mode (position mode with profile velocity)
 * Add passive operation mode
 * Fix velocity operation mode